Extract StatusBarView from Document.swift
Document.swift bundled the entire floating StatusBarView (~210 lines) after the
NSDocument subclass. Move it to its own StatusBarView.swift (pure code move),
leaving Document.swift focused on the document lifecycle (open/save/rename,
window setup, status-bar metric updates).

StatusBarView was widened from private to internal so Document can reference it
from the sibling file.
